# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/additional/insurances": {
            "get": {
                "tags": [
                    "Additional/Insurances"
                ],
                "summary": "Get list of insurances",
                "description": "Returns list of insurances",
                "operationId": "ed6f25c8814b0f930fa8f9ab2ad7097d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Insurance"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Additional/Insurances"
                ],
                "summary": "Store a new insurance",
                "description": "Store a new insurance",
                "operationId": "6ed68bf560d5a271011fa6d93f08a8a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsuranceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Insurance created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Insurance"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/api/additional/insurances/{insurance}": {
            "get": {
                "tags": [
                    "Additional/Insurances"
                ],
                "summary": "Get a specific insurance",
                "description": "Get a specific insurance by its ID",
                "operationId": "c88b6b2d7aab7357f5d48623e88988b5",
                "parameters": [
                    {
                        "name": "insurance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Insurance"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Additional/Insurances"
                ],
                "summary": "Update an existing insurance",
                "description": "Update an existing insurance by its ID",
                "operationId": "b3474046d0cc33c41eff380a9b7fe3fc",
                "parameters": [
                    {
                        "name": "insurance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsuranceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Insurance updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Insurance"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Additional/Insurances"
                ],
                "summary": "Delete an insurance",
                "description": "Delete an insurance by its ID",
                "operationId": "bcf1b0da08b5f0ec526ee0e7c869f583",
                "parameters": [
                    {
                        "name": "insurance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insurance deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InsuranceRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Insurance's name",
                        "type": "string",
                        "example": "ACME Supplies"
                    }
                },
                "type": "object"
            },
...
            "Insurance": {
                "required": [
                    "id",
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "Insurance's ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Insurance's name",
                        "type": "string",
                        "example": "ACME Supplies"
                    },
                    "created_at": {
                        "description": "Date and time of creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time of the last update",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}